# product-contract/openapi/sandbox-runtime-product-v1alpha1.yaml
openapi: 3.1.0
info:
  title: Sandbox Runtime Product API
  version: 0.1.0
  description: >-
    Phase 1 design authority for the authenticated Product control plane. This
    document is not an implementation or availability claim. Provider and
    local management APIs are separate surfaces.
  license:
    name: MIT
paths:
...
  /api/v1/workspaces/{workspace_id}/sessions:
    get:
      operationId: listWorkspaceSessions
      summary: List runtime sessions for one Workspace
      parameters:
        - $ref: '#/components/parameters/WorkspaceID'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Bounded session page
          content:
            application/json:
              schema:
                $ref: ../schemas/product-v1alpha1.schema.json#/$defs/SessionPage
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
    post:
      operationId: createRuntimeSession
      summary: Accept asynchronous creation of a Product runtime session
      x-max-body-bytes: 65536
      parameters:
        - $ref: '#/components/parameters/WorkspaceID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../schemas/product-v1alpha1.schema.json#/$defs/CreateSessionRequest
      responses:
        '202':
          description: Product operation accepted
          content:
            application/json:
              schema:
                $ref: ../schemas/product-v1alpha1.schema.json#/$defs/ProductOperation
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Unsupported'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /api/v1/sessions/{session_id}:
    get:
      operationId: getRuntimeSession
      summary: Read one Product runtime session
      parameters:
        - $ref: '#/components/parameters/SessionID'
      responses:
        '200':
          description: Product runtime session
          content:
            application/json:
              schema:
                $ref: ../schemas/product-v1alpha1.schema.json#/$defs/RuntimeSession
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/v1/sessions/{session_id}:close:
    post:
      operationId: closeRuntimeSession
      summary: Accept asynchronous Product session closure
      x-max-body-bytes: 32768
      parameters:
        - $ref: '#/components/parameters/SessionID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../schemas/product-v1alpha1.schema.json#/$defs/CloseSessionRequest
      responses:
        '202':
          description: Product operation accepted
          content:
            application/json:
              schema:
                $ref: ../schemas/product-v1alpha1.schema.json#/$defs/ProductOperation
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '503': {$ref: '#/components/responses/Unavailable'}
...
security:
  - bearerAuth: []
components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: '^[!-~]+$'
    Cursor:
      name: cursor
      in: query
      required: false
      schema: {type: string, minLength: 1, maxLength: 1024}
    Limit:
      name: limit
      in: query
      required: false
      schema: {type: integer, minimum: 1, maximum: 200, default: 50}
...
    WorkspaceID: {$ref: '#/components/parameters/WorkspaceIDValue'}
    WorkspaceIDValue:
      name: workspace_id
      in: path
      required: true
      schema: {$ref: ../schemas/product-v1alpha1.schema.json#/$defs/Identifier}
...
    SessionID:
      name: session_id
      in: path
      required: true
      schema: {$ref: ../schemas/product-v1alpha1.schema.json#/$defs/Identifier}
...
  responses:
    BadRequest:
      description: Request syntax, schema, bounds, or metadata is invalid
      content:
        application/json:
          schema: {$ref: ../schemas/product-v1alpha1.schema.json#/$defs/ProductError}
    Unauthorized:
      description: Product authentication failed
      content:
        application/json:
          schema: {$ref: ../schemas/product-v1alpha1.schema.json#/$defs/ProductError}
    Forbidden:
      description: Authenticated actor is not authorized
      content:
        application/json:
          schema: {$ref: ../schemas/product-v1alpha1.schema.json#/$defs/ProductError}
    NotFound:
      description: Authorized resource is not found
      content:
        application/json:
          schema: {$ref: ../schemas/product-v1alpha1.schema.json#/$defs/ProductError}
    Conflict:
      description: Version, idempotency, state, or control authority conflicts
      content:
        application/json:
          schema: {$ref: ../schemas/product-v1alpha1.schema.json#/$defs/ProductError}
...
    Unsupported:
      description: Capability or protocol is not currently advertised
      content:
        application/json:
          schema: {$ref: ../schemas/product-v1alpha1.schema.json#/$defs/ProductError}
    RateLimited:
      description: Product or Gateway capacity is temporarily exhausted
      headers:
        Retry-After:
          required: true
          schema: {type: integer, minimum: 1}
      content:
        application/json:
          schema: {$ref: ../schemas/product-v1alpha1.schema.json#/$defs/ProductError}
    Unavailable:
      description: Required Product dependency is temporarily unavailable
      headers:
        Retry-After:
          required: false
          schema: {type: integer, minimum: 1}
      content:
        application/json:
          schema: {$ref: ../schemas/product-v1alpha1.schema.json#/$defs/ProductError}
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: short-lived-access-token